pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "palinode"
version = "0.7.0"
description = "The memory substrate for AI agents and developer tools. Git-versioned, file-native, MCP-first."
authors = [
  {name = "Paul Kyle", email = "paul@phasespace.co"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai", "agents", "memory", "mcp", "llm", "rag", "knowledge-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "sqlite-vec",
    "watchdog",
    "pyyaml",
    "httpx",
    "python-frontmatter",
    "fastapi",
    "uvicorn",
    "pydantic",
    "mcp",
    "rich",
    "click",
]

[project.urls]
"Homepage" = "https://github.com/phasespace-labs/palinode"
"Documentation" = "https://github.com/phasespace-labs/palinode/tree/main/docs"
"Repository" = "https://github.com/phasespace-labs/palinode"
"Issues" = "https://github.com/phasespace-labs/palinode/issues"
"Changelog" = "https://github.com/phasespace-labs/palinode/blob/main/docs/CHANGELOG.md"
"Privacy" = "https://github.com/phasespace-labs/palinode/blob/main/PRIVACY.md"

[project.scripts]
palinode-watcher = "palinode.indexer.watcher:main"
palinode-api = "palinode.api.server:main"
palinode = "palinode.cli:main"
palinode-mcp = "palinode.mcp:main"
palinode-mcp-http = "palinode.mcp:main_http"
palinode-mcp-sse = "palinode.mcp:main_sse"  # deprecated alias for main_http

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
# Exclude live tests from default runs (they need a running server)
addopts = "--ignore=tests/live"
markers = [
    "slow: marks tests that need Ollama or take >5s",
]
asyncio_mode = "strict"

[tool.setuptools]
packages = ["palinode", "palinode.core", "palinode.indexer", "palinode.api", "palinode.ingest", "palinode.consolidation", "palinode.migration", "palinode.cli"]
